+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+# Customico CHANHE LOG
+
+## 0.0.8
+
+   Fixed some package bugs
+
+   improve the performance
+
+## 0.0.7
+
+## Added
+
+1- Phone format for CustomPhoneForm
+
+## Fixed
+
+1- Fixed package bugs
+
+## 0.0.6
+
+bug fixes
+
+## 0.0.5
+
+bug fixes
+
+## 0.0.4
+
+bug fixes
+
+## 0.0.3
+
+bug fixes
+
+## 0.0.2
+
+bug fixes
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+Package release